requirements.txt
pandas==1.3.4
python-binance==1.0.12